refactor: add missing flights prefix for booking route segments
