+++ src/app/(frontend)/(without-sidebar)/flights/booking/confirmation/page.tsx
