+++ src/app/(frontend)/(without-sidebar)/flights/booking/confirmation/queries.ts
