+++ src/app/(frontend)/(without-sidebar)/flights/booking/payment/page.tsx
